requirements.txt
argparse
cchardet>=1.1.2
certifi>=2017.4.17
chardet<3.1.0,>=3.0.2
cssselect
feedparser==5.2.1
geonamescache==1.3.0
html5lib
lxml~=4.0
python-dateutil==2.8.2
requests>=2.20.0
selenium==4.0.0-alpha-6
tldextract
